% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_ancova.R
\name{n_ancova}
\alias{n_ancova}
\title{Sample Size Calculation for the Analysis of Covariances (ANCOVA)}
\usage{
n_ancova(alpha, power, r, r.strict = TRUE, effect, corr, sd, gs = TRUE)
}
\arguments{
\item{alpha}{Significance level \eqn{\alpha}.}

\item{power}{Desired Power \eqn{1-\beta}.}

\item{r}{Quotient of Sample sizes \eqn{r = n_Y / n_X}.}

\item{r.strict}{Allocation Handling. Default = TRUE. If set to TRUE,
the Sample size  per Group is round up to the next combination hitting the
specified allocation exactly. If set to FALSE, the Sample size per
Group is round up to the next natural number.}

\item{effect}{Effect \eqn{\Delta_A} used as alternative hypothesis.}

\item{corr}{Correlation  \eqn{rho} to Covariate.}

\item{sd}{Standard deviation \eqn{\sigma}.}

\item{gs}{Guenther/Schouten correcture. Default = TRUE. If set to TRUE,
the G/S correcture like in (4.15) in [1] is used.}
}
\value{
\code{n_ancova} returns an object of type list. The resulting
  Sample Sizes are located in entrys named \code{n_X}, \code{n_Y}, \code{n}.
  The resulting power is named \code{power_out}.
  WARNING: This function is not implemented yet.
}
\description{
\code{n_ancova} performs the Sample Size calculation for the test on
  mean difference for two samples with an Analysis of
  Covariances (ANCOVA).
  The method is based on the pages 18 - 20 in [1]. The Sample Size
  is calculated using an approximative formula,
  suggested by Frison & Pocock (1999), implemented (4.14) in [1].
}
\examples{
n_ancova(alpha = .05, power = .90, r = 1, effect = 10, corr = .5, sd = 20)$n

}
