% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_chisq_mult_groups.R
\name{n_chisq_mult_groups}
\alias{n_chisq_mult_groups}
\title{Sample Size Calculation for the Chi-Square Test comparing rates of
\eqn{k > 2} groups}
\usage{
n_chisq_mult_groups(alpha, power, n.groups, p_A)
}
\arguments{
\item{alpha}{Significance level \eqn{\alpha}.}

\item{power}{Desired Power \eqn{1-\beta}.}

\item{n.groups}{Number of Groups \eqn{k}.}

\item{p_A}{Vector \eqn{\p_A} of expected rates on the alternative.}
}
\value{
\code{n_chisq_mult_groups} returns an object of type list. The resulting
  Sample Sizes are located in entrys named \code{n_per_group}, \code{n}.
  The resulting power is named \code{power_out}.
}
\description{
\code{n_chisq_mult_groups} performs the Sample Size calculation for
  the Chi-Square test comparing rates of \eqn{k > 2} independent samples.
  The method used here is based on the page 30 in [1].
  The Sample Size is calculated using an iterative approach using (6.3) in [1].
\describe{
  \item{Null Hypothesis:}{\eqn{p_1 = p_2 = ... = p_k}}
  \item{Alternative Hypothesis:}{There are \eqn{i,j} with \eqn{p_i \ne p_j}}
}
}
\details{
[1] M.Kieser: Fallzahlberechnung in der medizinischen Forschung [2018], 1th Edition
}
\examples{
n_chisq_mult_groups(alpha = .05, power = .8, n.groups = 3, p_A = c(.3, .5, .4))

}
