% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_fisher_boschloo.R
\name{n_fisher_boschloo}
\alias{n_fisher_boschloo}
\title{Sample Size Calculation for the Fisher-Boschloo-Test}
\usage{
n_fisher_boschloo(p_Y, p_X, alpha, power, r, exact = TRUE, SW = 1e-04,
  MAXH = 10)
}
\arguments{
\item{p_Y}{Event rate of Group Y on the alternative.}

\item{p_X}{Event rate of Group X on the alternative.}

\item{alpha}{Significance level \eqn{\alpha}.}

\item{power}{Desired Power \eqn{1-\beta}.}

\item{r}{Quotient of Sample sizes \eqn{r = n_Y / n_X}.}

\item{exact}{Default = TRUE. If set to FALSE the approximative
sample size is calculated. On TRUE the exact sample size is calculated.}

\item{SW}{Default = .0001. Step width.}

\item{MAXH}{Default = 10.}
}
\value{
\code{n_fisher_boschloo} returns an object of type list. The resulting
  Sample Sizes are located in entrys named \code{n_X}, \code{n_Y}, \code{n}.
  The resulting power is named \code{power_out}.
}
\description{
\code{n_fisher_boschloo} performs the Sample Size calculation for two
  independent samples with binary data using the absolute rate
  difference quantifying the effect of an intervention.
  The method used here is written by S. Wellek.
  See [2] for further details.
}
\details{
\describe{
  \item{Null Hypothesis:}{\eqn{p_Y - p_X = 0}}
  \item{Alternative Hypothesis:}{\eqn{|p_Y - p_X| \ge \Delta_A}}
}

[2] S. Wellek: Nearly exact sample size calculation for powerful
non-randomized tests for differences between binomial proportions [2015],
Statistica Neerlandica
}
\examples{
n_fisher_boschloo(p_Y = .5, p_X = .3, alpha = .05, power = .8, r = 2)
n_fisher_boschloo(p_Y = .5, p_X = .3, alpha = .05, power = .8, r = 2)$n
n_fisher_boschloo(p_Y = .5, p_X = .3, alpha = .05, power = .8, r = 2, power.exact = FALSE)

}
