% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_ttest.R
\name{n_ttest}
\alias{n_ttest}
\title{Sample Size Calculation for the two-sided t-Test}
\usage{
n_ttest(effect, sd, alpha, power, r = 1)
}
\arguments{
\item{effect}{Effect \eqn{\Delta_A} used as alternative hypothesis.}

\item{sd}{Standard deviation \eqn{\sigma} of the Data on the
alternative hypothesis.}

\item{alpha}{Significance level \eqn{\alpha}.}

\item{power}{Desired Power \eqn{1-\beta}.}

\item{r}{Default = 1. Quotient of Group sizes \eqn{r = n_Y / n_X}.}
}
\value{
\code{n_ttest} returns an object of type list. The resulting
  Sample Sizes are located in entrys named \code{n_X}, \code{n_Y}, \code{n}.
  The resulting power is named \code{power_out}.
}
\description{
\code{n_ttest} performs the Sample Size calculation for the t-Test
  to the level \eqn{\alpha} comparing two independent samples.
  The method is based on the pages 16 - 18 in [1].
\describe{
  \item{Null Hypothesis:}{\eqn{\mu_Y - \mu_X = 0}}
  \item{Alternative Hypothesis:}{\eqn{|\mu_Y - \mu_X| \ge \Delta_A}}
}
}
\details{
[1] M.Kieser: Fallzahlberechnung in der medizinischen Forschung [2018], 1th Edition
}
\examples{
n_ttest(effect = 10, sd = 20, alpha = .05, power = .90)
n_ttest(effect = 10, sd = 20, alpha = .05, power = .90, r = 2)$n
n_ttest(effect = 10, sd = 20, alpha = .05, power = .90, r = 2)$power_out

}
