% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_chisq.R
\name{power_binomial}
\alias{power_binomial}
\title{Power Calculation for the Chi-Square Test}
\usage{
power_binomial(p_Y, p_X, n_Y, n_X, alpha, power.exact)
}
\arguments{
\item{p_Y}{Event rate of group Y on the alternative.}

\item{p_X}{Event rate of group X on the alternative.}

\item{n_Y}{Sample size of group Y.}

\item{n_X}{Sample size of group X.}

\item{alpha}{Significance level \eqn{\alpha}.}

\item{power.exact}{If set to FALSE an approximative distributive
is used for calculating the power, given the alternative distribution at the
bottom of p. 22 in [1]. On TRUE
the iterative approach is used.}
}
\value{
\code{power_binomial} returns the power.
}
\description{
\code{power_chisq} performs the power calculation for two
  independent samples with binary data using the absolute rate
  difference quantifying the effect of an intervention.
  The method used here is based on the pages 21 - 26 in [1].
}
\details{
[1] M.Kieser: Fallzahlberechnung in der medizinischen Forschung [2018], 1th Edition
}
\examples{
power_binomial(p_Y = .5, p_X = .3, n_Y = 100, n_X = 50, alpha = .05, power.exact = TRUE)

}
