% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_chisq_mult_groups.R
\name{power_chisq_mult_groups}
\alias{power_chisq_mult_groups}
\title{Power Calculation for the Chi-Square Test comparing rates of
\eqn{k > 2} groups}
\usage{
power_chisq_mult_groups(p_A, n.groups, n_per_group, alpha)
}
\arguments{
\item{p_A}{Vector \eqn{\p_A} of expected rates on the alternative.}

\item{n.groups}{Number of Groups \eqn{k}.}

\item{n_per_group}{Sample size per group.}

\item{alpha}{Significance level \eqn{\alpha}.}
}
\value{
\code{power_chisq_mult_groups} returns the power.
}
\description{
\code{power_chisq_mult_groups} performs the power calculation for
  the Chi-Square test comparing rates of \eqn{k > 2} independent samples.
  The method used here is based on the page 30 in [1].
  The power is calculated using (6.2) in [1].
}
\details{
[1] M.Kieser: Fallzahlberechnung in der medizinischen Forschung [2018], 1th Edition
}
\examples{
power_chisq_mult_groups(p_A = c(.3, .5, .4), n.groups = 3, n_per_group = 70, alpha = .05)

}
