% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_ftest.R
\name{power_ftest}
\alias{power_ftest}
\title{Power Calculation for the f Test comparing means of \eqn{k > 2} groups}
\usage{
power_ftest(mu_A, sd, n.groups, n_per_group, alpha)
}
\arguments{
\item{mu_A}{Vector \eqn{\mu_A} of expected means on the alternative.}

\item{sd}{Standard deviation \eqn{sigma}.}

\item{n.groups}{Number of Groups \eqn{k}.}

\item{n_per_group}{Sample size per group.}

\item{alpha}{Significance level \eqn{\alpha}.}
}
\value{
\code{n_ftest} returns the power.
}
\description{
\code{power_ftest} performs the power calculation for the f-test
  comparing means of \eqn{k > 2} independent samples.
  The method used here is based on the page 29 in [1].
}
\details{
[1] M.Kieser: Fallzahlberechnung in der medizinischen Forschung [2018], 1th Edition
}
\examples{
power_ftest(mu_A = c(1, 2, 3), sd = 20, n.groups = 3, n_per_group = 70, alpha = .025)

}
