% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_ztest.R
\name{power_ztest}
\alias{power_ztest}
\title{Power calculation for the two-sided z-Test}
\usage{
power_ztest(effect, sd, n_X, alpha, r = 1)
}
\arguments{
\item{effect}{Effect \eqn{\Delta_A} used as alternative hypothesis.}

\item{sd}{Standard deviation \eqn{\sigma}.}

\item{n_X}{Sample size of group X.}

\item{alpha}{Significance level \eqn{\alpha}.}

\item{r}{Default = 1. Quotient of Group sizes \eqn{r = n_Y / n_X}.}
}
\value{
\code{n_ztest} returns the power.
}
\description{
\code{n_ztest} performs the power calculation for the z-Test
  to the level \eqn{\alpha} comparing two independent samples.
  The method is based on the pages 13 - 16 in [1].
}
\details{
[1] M.Kieser: Fallzahlberechnung in der medizinischen Forschung [2018], 1th Edition
}
\examples{
power_ztest(effect = 10, sd = 20, n_X = 70, alpha = .05)

}
