% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesizr.R
\docType{package}
\name{samplesizr}
\alias{samplesizr}
\alias{samplesizr-package}
\title{samplesizr: Sample size calculations for [Buchtitel]}
\description{
samplesizr: Sample size calculations for [Buchtitel]
}
\section{Functions}{

\itemize{

  \item{\code{\link{n_chisqr}}: }{bla}

  \item{\code{\link{n_chisq_mult_groups}}: }{blubb}

}
}

