% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_z_test.R
\name{n_ztest}
\alias{n_ztest}
\title{Sample size for z-Test}
\usage{
n_ztest(sig, pow, r, r.strict = TRUE, effect, std)
}
\arguments{
\item{sig}{Significance level}

\item{pow}{Power}

\item{r}{Allocation}

\item{std}{Std. deviation}
}
\value{
Sample size for Group 1
}
\description{
Sample size for z-Test
}
