% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_ztest.R
\name{n_ztest}
\alias{n_ztest}
\title{Sample Size Calculation for the two-sided z-Test}
\usage{
n_ztest(effect, sd, alpha, power, r)
}
\arguments{
\item{effect}{Effect \eqn{\Delta_A} used as alternative hypothesis.}

\item{sd}{Standard deviation \eqn{\sigma}.}

\item{alpha}{Significance level \eqn{\alpha}.}

\item{power}{Desired Power \eqn{1-\beta}.}

\item{r}{Quotient of Group sizes \eqn{r = n_Y / n_X}.}
}
\value{
\code{n_ztest} returns an object of type list. The resulting
  Sample Sizes are located in entrys named \code{n_X}, \code{n_Y}, \code{n}.
  The resulting power is named \code{power_out}
}
\description{
\code{n_ztest} performs the Sample Size calculation for the z-Test
  to the level \eqn{\alpha} comparing two independent samples.
  The method is based on the pages 13 - 16 in [1].
\describe{
  \item{Null Hypothesis:}{\eqn{\mu_Y - \mu_X = 0}}
  \item{Alternative Hypothesis:}{\eqn{\mu_Y - \mu_X \ge \Delta_A}}
}
}
\details{
[1] M.Kieser: Fallzahlberechnung in der medizinischen Forschung [2018], 1th Edition
}
\examples{
n_ztest(effect = .5, sd = 1, alpha = .05, power = .90, r = 2)
n_ztest(effect = .5, sd = 1, alpha = .05, power = .90, r = 2)$n
n_ztest(effect = .5, sd = 1, alpha = .05, power = .90, r = 2)$power_out

}
